Normalize current TCVT tests

+++ test/tilelang_st/npu/a5/src/st/testcase/tcvt/launch.cpp
@@ -15,8 +15,8 @@
 extern "C" __global__ AICORE void TCVT_f32_to_i32_rint_16x64(__gm__ float *src, __gm__ int32_t *dst);
 extern "C" __global__ AICORE void TCVT_f32_to_i32_round_16x64(__gm__ float *src, __gm__ int32_t *dst);
 extern "C" __global__ AICORE void TCVT_i32_to_f32_rint_16x64(__gm__ int32_t *src, __gm__ float *dst);
-extern "C" __global__ AICORE void TCVT_f32_to_f16_rint_16x64(__gm__ float *src, __gm__ half *dst);
-extern "C" __global__ AICORE void TCVT_f16_to_f32_rint_16x64(__gm__ half *src, __gm__ float *dst);
+extern "C" __global__ AICORE void TCVT_f32_to_f16_rint_16x64(__gm__ float *src, __gm__ uint16_t *dst);
+extern "C" __global__ AICORE void TCVT_f16_to_f32_rint_16x64(__gm__ uint16_t *src, __gm__ float *dst);
 
 void LaunchTCVT_f32_to_i32_rint_16x64(void *src, void *dst, void *stream) {
     TCVT_f32_to_i32_rint_16x64<<<1, nullptr, stream>>>((__gm__ float *)src, (__gm__ int32_t *)dst);
@@ -31,9 +31,9 @@ void LaunchTCVT_i32_to_f32_rint_16x64(void *src, void *dst, void *stream) {
 }
 
 void LaunchTCVT_f32_to_f16_rint_16x64(void *src, void *dst, void *stream) {
-    TCVT_f32_to_f16_rint_16x64<<<1, nullptr, stream>>>((__gm__ float *)src, (__gm__ half *)dst);
+    TCVT_f32_to_f16_rint_16x64<<<1, nullptr, stream>>>((__gm__ float *)src, (__gm__ uint16_t *)dst);
 }
 
 void LaunchTCVT_f16_to_f32_rint_16x64(void *src, void *dst, void *stream) {
-    TCVT_f16_to_f32_rint_16x64<<<1, nullptr, stream>>>((__gm__ half *)src, (__gm__ float *)dst);
+    TCVT_f16_to_f32_rint_16x64<<<1, nullptr, stream>>>((__gm__ uint16_t *)src, (__gm__ float *)dst);
 }
